LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/transition-style-change-event-expected.txt

Green

FAIL Container Queries - Style Change Event for transitions assert_equals: @container queries supported expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
